# Conflicts:
#	python/paddle/incubate/sparse/__init__.py
#	python/paddle/utils/code_gen/sparse_bw_api.yaml

+++ paddle/fluid/distributed/ps/service/brpc_ps_client.cc
@@ -197,7 +197,7 @@ int32_t BrpcPsClient::Initialize() {
 
   // 异步push 请求队列初始化
   const auto &worker_param = _config.worker_param().downpour_worker_param();
-  for (size_t i = 0; i < worker_param.downpour_table_param_size(); ++i) {
+  for (int i = 0; i < worker_param.downpour_table_param_size(); ++i) {
     auto type = worker_param.downpour_table_param(i).type();
     auto table_id = worker_param.downpour_table_param(i).table_id();
     if (type == PS_DENSE_TABLE) {
@@ -662,7 +662,7 @@ std::future<int32_t> BrpcPsClient::PushSparseParam(size_t table_id,
     char *push_data_ptr = const_cast<char *>(push_data->data());
     memcpy(push_data_ptr, kvs.data(), kv_size * sizeof(uint64_t));
     push_data_ptr += kv_size * sizeof(uint64_t);
-    for (int i = 0; i < kv_size; ++i) {
+    for (size_t i = 0; i < kv_size; ++i) {
       memcpy(push_data_ptr, value_ptr[i], value_size);
       push_data_ptr += value_size;
     }
@@ -882,7 +882,7 @@ std::future<int32_t> BrpcPsClient::PushSparseRawGradient(
     memcpy(push_data_ptr, kvs.data(), kv_size * sizeof(uint64_t));
     push_data_ptr += kv_size * sizeof(uint64_t);
 
-    for (int i = 0; i < kv_size; ++i) {
+    for (size_t i = 0; i < kv_size; ++i) {
       memcpy(push_data_ptr, value_ptr[i], value_size);
       push_data_ptr += value_size;
     }
@@ -1237,7 +1237,7 @@ std::future<int32_t> BrpcPsClient::PushSparseRawGradientPartial(
   char *push_data_ptr = const_cast<char *>(push_data->data());
   memcpy(push_data_ptr, keys, num * sizeof(uint64_t));
   push_data_ptr += num * sizeof(uint64_t);
-  for (int i = 0; i < num; ++i) {
+  for (uint32_t i = 0; i < num; ++i) {
     memcpy(push_data_ptr, update_values[i], value_size);
     push_data_ptr += value_size;
   }
@@ -1257,7 +1257,7 @@ int32_t BrpcPsClient::RecvAndSaveTable(const uint64_t table_id,
   int64_t var_shape = 0;
   std::string table_class;
   const auto &worker_param = _config.worker_param().downpour_worker_param();
-  for (size_t i = 0; i < worker_param.downpour_table_param_size(); ++i) {
+  for (int i = 0; i < worker_param.downpour_table_param_size(); ++i) {
     if (worker_param.downpour_table_param(i).table_id() == table_id) {
       var_name = worker_param.downpour_table_param(i).common().table_name();
       var_num = worker_param.downpour_table_param(i).common().table_num();
@@ -1481,13 +1481,13 @@ void BrpcPsClient::PushSparseTaskConsume() {
         closure->add_timer(rpc_timer);
 
         std::vector<std::future<int>> merge_status(request_call_num);
-        for (int shard_idx = 0; shard_idx < request_call_num; ++shard_idx) {
+        for (size_t shard_idx = 0; shard_idx < request_call_num; ++shard_idx) {
           merge_status[shard_idx] =
               async_push_sparse_shard_threads.enqueue(std::bind(
                   &BrpcPsClient::PushSparseAsyncShardPush, this, task_list,
                   request_kv_num, table_id, shard_idx, closure, accessor));
         }
-        for (int shard_idx = 0; shard_idx < request_call_num; ++shard_idx) {
+        for (size_t shard_idx = 0; shard_idx < request_call_num; ++shard_idx) {
           merge_status[shard_idx].wait();
         }
         merge_status.clear();
@@ -1497,13 +1497,13 @@ void BrpcPsClient::PushSparseTaskConsume() {
         auto queue_size = task_queue->Size();
       } else {  // 未达到阈值 只做多路归并
         std::vector<std::future<int>> merge_status(request_call_num);
-        for (int shard_idx = 0; shard_idx < request_call_num; ++shard_idx) {
+        for (size_t shard_idx = 0; shard_idx < request_call_num; ++shard_idx) {
           merge_status[shard_idx] =
               async_push_sparse_shard_threads.enqueue(std::bind(
                   &BrpcPsClient::PushSparseAsyncShardMerge, this, task_list,
                   request_kv_num, table_id, shard_idx, accessor));
         }
-        for (int shard_idx = 0; shard_idx < request_call_num; ++shard_idx) {
+        for (size_t shard_idx = 0; shard_idx < request_call_num; ++shard_idx) {
           merge_status[shard_idx].wait();
         }
 
@@ -1529,7 +1529,7 @@ void sparse_local_merge(ValueAccessor *accessor, float *merge_data,
   size_t col_num = accessor->GetAccessorInfo().update_dim;
   float *merge_data_shell[col_num];
   const float *another_data_shell[col_num];
-  for (int i = 0; i < col_num; ++i) {
+  for (size_t i = 0; i < col_num; ++i) {
     merge_data_shell[i] = merge_data + i;
     another_data_shell[i] = another_data + i;
   }
@@ -1546,12 +1546,12 @@ int BrpcPsClient::PushSparseAsyncShardMerge(
 
   thread_local std::vector<std::pair<uint64_t, const float *>> sorted_kv_list;
   sorted_kv_list.clear();
-  for (int i = 1; i < task_list.size(); ++i) {
+  for (size_t i = 1; i < task_list.size(); ++i) {
     size_t kv_num = task_list[i]->data()->shared_data[shard_idx].kv_num;
     auto &key_list = task_list[i]->data()->shared_data[shard_idx].key_list;
     auto &value_list = task_list[i]->data()->shared_data[shard_idx].value_list;
 
-    for (int j = 0; j < kv_num; ++j) {
+    for (size_t j = 0; j < kv_num; ++j) {
       if (value_list[j].size() < value_size) {
         LOG(WARNING) << "value_list[" << j << "]: " << value_list[j].c_str()
                      << "is invalid.";
@@ -1654,7 +1654,7 @@ int BrpcPsClient::PushSparseAsyncShardPush(
   memcpy(push_data_ptr, merged_key_list.data(),
          merged_kv_count * sizeof(uint64_t));
   push_data_ptr += merged_kv_count * sizeof(uint64_t);
-  for (int i = 0; i < merged_kv_count; ++i) {
+  for (size_t i = 0; i < merged_kv_count; ++i) {
     const char *task_data_ptr = merged_value_list[i].data();
 
     memcpy(push_data_ptr, (float *)(task_data_ptr),  // NOLINT
@@ -1778,7 +1778,7 @@ void BrpcPsClient::PushDenseTaskConsume() {
               });
           ++merge_count;
         }
-        for (int i = 0; i < merge_count; ++i) {
+        for (uint32_t i = 0; i < merge_count; ++i) {
           merge_status[i].wait();
         }
